requirements.txt
streamlit==1.28.1
openai==1.3.5
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
speechrecognition==3.10.0
pydub==0.25.1
gtts==2.4.0
pygame==2.5.2
pyaudio==0.2.11
numpy==1.24.3
pandas==1.5.3
plotly==5.17.0
streamlit-audio-recorder==0.0.8
